pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trend-spotter"
version = "0.1.2"
description = "AI-powered trend analysis tool using Google ADK"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "google-ai-generativelanguage",
    "google-generativeai", 
    "google-cloud-aiplatform",
    "google-adk",
    "python-dotenv",
    "streamlit",
    "pandas",
    "plotly",
    "requests",
    "beautifulsoup4",
    "lxml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
    "bandit",
    "safety",
]

[project.urls]
"Homepage" = "https://github.com/subhashishd/trend-spotter"
"Bug Reports" = "https://github.com/subhashishd/trend-spotter/issues"
"Source" = "https://github.com/subhashishd/trend-spotter"

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
profile = "black"
line_length = 79
src_paths = ["."]

[tool.mypy]
# Ignore missing imports for third-party libraries
ignore_missing_imports = true
# Don't require type annotations for all functions
no_strict_optional = true
# Show error codes
show_error_codes = true
# Use explicit package bases to avoid module name conflicts
explicit_package_bases = true
# Exclude virtual environment and other non-source directories
exclude = [
    "myenv/",
    "build/",
    "dist/",
    ".mypy_cache/",
    ".git/",
    "trend_spotter.egg-info/",
]
[tool.pytest.ini_options]
# Pytest configuration
testpaths = ["tests"]
# Add markers for different test types
markers = [
    "unit: Unit tests that don't require external services",
    "integration: Integration tests that require Google Cloud credentials",
    "slow: Slow tests that may take a long time",
]
# Configure asyncio for async tests
asyncio_mode = "auto"
# Test configuration
addopts = [
    "--tb=short",
    "--strict-markers",
    "-v",
]
# Patterns to ignore during collection
norecursedirs = [
    "myenv",
    "build",
    "dist",
    ".git",
    "*.egg",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["trend_spotter*"]
exclude = ["myenv*", "tests*", "*.egg-info*", "build*", "dist*"]

[tool.adk.agents]
trend_spotter = "trend_spotter.agent:root_agent"
